Big overhaul, reverting back to 'wrapping' element
Really a 'view' is just another model. We'd like to create a new model,
just for a view. However it would be a waste to duplicate the elements
themselves. Thus the desire to wrap.

Of course, in the future there will probably be a way to optimise
better, using the model as is, and instead a record of which element to
ignore in the view. But I believe this will be the simplest solution for
now