Change yaml to be strict and remove unused config flag
Now that the config is part of the architecture yaml file the config
flag is no longer used anywhere.

Additionally, if a invalid yaml was submitted the error was opaque. I
somewhat improved that by making yaml unmarshaling strict and therefore
returning an error if yaml structure does not match.